streamlit_lite.py
#!/usr/bin/env python3
"""
Lightweight Streamlit app for Ocean AI QA Framework
Optimized for Render deployment without heavy ML dependencies
"""

import streamlit as st
import os
import json
import tempfile
import subprocess
from pathlib import Path
from typing import List, Dict, Any, Optional
import pandas as pd
import requests
from datetime import datetime

# Load environment variables
from dotenv import load_dotenv
load_dotenv()

# Import Google Gemini AI (lightweight)
try:
    import google.generativeai as genai
    GEMINI_AVAILABLE = True
    api_key = os.getenv('GEMINI_API_KEY')
    if api_key and api_key != 'your_gemini_api_key_here':
        genai.configure(api_key=api_key)
        GEMINI_CONFIGURED = True
    else:
        GEMINI_CONFIGURED = False
except ImportError:
    GEMINI_AVAILABLE = False
    GEMINI_CONFIGURED = False

class SimpleTestGenerator:
    """Lightweight test case generator without heavy ML dependencies"""
    
    def __init__(self):
        self.templates = {
            'discount_code': [
                {
                    'test_id': 'TC-001',
                    'feature': 'Discount Code',
                    'scenario': 'Apply valid discount code SAVE15',
                    'steps': ['Navigate to checkout', 'Enter SAVE15 code', 'Click Apply'],
                    'expected': '15% discount applied to total'
                },
                {
                    'test_id': 'TC-002', 
                    'feature': 'Discount Code',
                    'scenario': 'Apply invalid discount code',
                    'steps': ['Navigate to checkout', 'Enter INVALID code', 'Click Apply'],
                    'expected': 'Error message displayed'
                }
            ],
            'form_validation': [
                {
                    'test_id': 'TC-003',
                    'feature': 'Form Validation',
                    'scenario': 'Submit form with empty required fields',
                    'steps': ['Leave name field empty', 'Click submit'],
                    'expected': 'Validation error for required field'
                }
            ],
            'cart_management': [
                {
                    'test_id': 'TC-004',
                    'feature': 'Cart Management', 
                    'scenario': 'Add item to cart',
                    'steps': ['Click add to cart button', 'Verify item added'],
                    'expected': 'Item appears in cart with quantity 1'
                }
            ]
        }
    
    def generate_test_cases(self, feature_type='all'):
        """Generate test cases for specified feature"""
        if feature_type == 'all':
            all_tests = []
            for tests in self.templates.values():
                all_tests.extend(tests)
            return all_tests
        
        return self.templates.get(feature_type, [])

def generate_selenium_script(test_case):
    """Generate Selenium automation script"""
    script = f'''
from selenium import webdriver
from selenium.webdriver.common.by import By
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
import time

class Test{test_case['test_id'].replace('-', '_')}:
    """
    Test Case: {test_case['test_id']}
    Feature: {test_case['feature']}
    Scenario: {test_case['scenario']}
    """
    
    def __init__(self):
        options = webdriver.ChromeOptions()
        options.add_argument("--headless")
        options.add_argument("--no-sandbox")
        options.add_argument("--disable-dev-shm-usage")
        self.driver = webdriver.Chrome(options=options)
        self.wait = WebDriverWait(self.driver, 10)
    
    def test_{test_case['test_id'].lower().replace('-', '_')}(self):
        try:
            # Navigate to application
            self.driver.get("https://your-app-url.com/checkout.html")
            
            # Test steps:
'''
    
    for i, step in enumerate(test_case['steps'], 1):
        script += f'            # Step {i}: {step}\n'
        script += f'            # TODO: Implement step "{step}"\n'
    
    script += f'''            
            # Verify expected result: {test_case['expected']}
            # TODO: Add assertions for expected result
            
            print("✅ Test passed: {test_case['test_id']}")
            
        except Exception as e:
            print(f"❌ Test failed: {test_case['test_id']} - {{e}}")
            raise
        finally:
            self.driver.quit()

if __name__ == "__main__":
    test = Test{test_case['test_id'].replace('-', '_')}()
    test.test_{test_case['test_id'].lower().replace('-', '_')}()
'''
    
    return script

def main():
    st.set_page_config(
        page_title="Ocean AI QA Framework",
        page_icon="🌊",
        layout="wide"
    )
    
    st.title("🌊 Ocean AI - Autonomous QA Framework")
    st.caption("Lightweight production version optimized for Render deployment")
    
    # Sidebar navigation
    st.sidebar.title("Navigation")
    page = st.sidebar.selectbox("Choose a page:", [
        "🏠 Home",
        "🧪 Generate Test Cases", 
        "⚡ Selenium Scripts",
        "📊 Test Results"
    ])
    
    if page == "🏠 Home":
        st.markdown("## Welcome to Ocean AI QA Framework")
        
        col1, col2 = st.columns(2)
        
        with col1:
            st.markdown("### 🚀 Features")
            st.markdown("""
            - **Smart Test Generation**: AI-powered test case creation
            - **Selenium Automation**: Auto-generated test scripts  
            - **Multiple Test Types**: Functional, UI, API testing
            - **Export Options**: Download test cases and scripts
            """)
        
        with col2:
            st.markdown("### 📈 Status")
            if GEMINI_CONFIGURED:
                st.success("✅ AI Integration: Active")
            else:
                st.info("🔧 AI Integration: Configure GEMINI_API_KEY")
            
            st.info("🌐 Deployment: Render Production")
            st.success("⚡ Mode: Lightweight")
    
    elif page == "🧪 Generate Test Cases":
        st.markdown("## 🧪 Test Case Generation")
        
        generator = SimpleTestGenerator()
        
        feature_type = st.selectbox(
            "Select feature to test:",
            ["all", "discount_code", "form_validation", "cart_management"]
        )
        
        if st.button("Generate Test Cases"):
            with st.spinner("Generating test cases..."):
                test_cases = generator.generate_test_cases(feature_type)
                
                st.success(f"✅ Generated {len(test_cases)} test cases")
                
                # Display test cases
                for test in test_cases:
                    with st.expander(f"📋 {test['test_id']}: {test['scenario']}"):
                        col1, col2 = st.columns(2)
                        
                        with col1:
                            st.markdown(f"**Feature:** {test['feature']}")
                            st.markdown(f"**Test ID:** {test['test_id']}")
                            st.markdown(f"**Scenario:** {test['scenario']}")
                        
                        with col2:
                            st.markdown("**Steps:**")
                            for i, step in enumerate(test['steps'], 1):
                                st.markdown(f"{i}. {step}")
                            
                            st.markdown(f"**Expected:** {test['expected']}")
                
                # Download option
                st.download_button(
                    "📥 Download Test Cases (JSON)",
                    data=json.dumps(test_cases, indent=2),
                    file_name=f"test_cases_{feature_type}_{datetime.now().strftime('%Y%m%d_%H%M%S')}.json",
                    mime="application/json"
                )
    
    elif page == "⚡ Selenium Scripts":
        st.markdown("## ⚡ Selenium Script Generation")
        
        # Sample test case for demo
        sample_test = {
            'test_id': 'TC-DEMO',
            'feature': 'Demo Feature',
            'scenario': 'Sample test scenario',
            'steps': ['Open application', 'Perform action', 'Verify result'],
            'expected': 'Test passes successfully'
        }
        
        st.markdown("### Generate Automation Script")
        
        if st.button("Generate Selenium Script"):
            with st.spinner("Creating automation script..."):
                script = generate_selenium_script(sample_test)
                
                st.success("✅ Selenium script generated!")
                
                st.code(script, language="python")
                
                # Download option
                st.download_button(
                    "📥 Download Script (.py)",
                    data=script,
                    file_name=f"test_{sample_test['test_id'].lower()}.py",
                    mime="text/plain"
                )
    
    elif page == "📊 Test Results":
        st.markdown("## 📊 Test Results Dashboard")
        
        # Mock test results
        results_data = {
            'Test ID': ['TC-001', 'TC-002', 'TC-003', 'TC-004'],
            'Feature': ['Discount Code', 'Discount Code', 'Form Validation', 'Cart Management'],
            'Status': ['✅ PASS', '❌ FAIL', '✅ PASS', '⏳ PENDING'],
            'Duration': ['2.3s', '1.8s', '3.1s', '-']
        }
        
        df = pd.DataFrame(results_data)
        st.dataframe(df, use_container_width=True)
        
        # Summary metrics
        col1, col2, col3, col4 = st.columns(4)
        
        with col1:
            st.metric("Total Tests", "4")
        with col2:
            st.metric("Passed", "2", delta="50%")
        with col3:
            st.metric("Failed", "1", delta="-25%")
        with col4:
            st.metric("Pending", "1")
    
    # Footer
    st.sidebar.markdown("---")
    st.sidebar.markdown("🌊 **Ocean AI QA Framework**")
    st.sidebar.markdown("Production deployment on Render")
    st.sidebar.markdown(f"Version: 1.0.0 | {datetime.now().strftime('%Y-%m-%d')}")

if __name__ == "__main__":
    main()